requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
sqlalchemy>=1.4.23,<2.0.0
pydantic>=1.8.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-dotenv>=0.19.0
httpx>=0.23.0
rasa==3.6.2
numpy>=1.22.0,<1.24.0
google-cloud-speech>=2.21.0
google-cloud-texttospeech>=2.14.1
SpeechRecognition==3.10.0
gTTS>=2.3.2
openai-whisper==20231117
torch>=2.0.0
stripe==8.4.0
websockets==10.4
pytest==8.0.1
certifi>=2024.2.2 